requirements.txt
annoy
chardet
datasketch
docopt
jedi
libcst
msgpack
opentelemetry-api==0.17b0
opentelemetry-exporter-jaeger==0.17b0
opentelemetry-exporter-prometheus==0.17b0
opentelemetry-sdk==0.17b0
prometheus-client
ptgnn
pystache
pyyaml
pyzmq
tqdm
typing_extensions
